+++ molo/forms/tests/test_templatetags.py
@@ -7,11 +7,12 @@ from django.contrib.sessions.middleware import SessionMiddleware
 from molo.core.models import Main, Languages, SiteLanguageRelation
 from molo.core.tests.base import MoloTestCaseMixin
 from molo.forms.models import (
-    MoloFormPage, MoloFormField,
+    MoloFormPage, MoloFormField, ArticlePageForms,
     FormsIndexPage, PersonalisableForm, MoloFormSubmission)
 
 from molo.forms.templatetags.molo_forms_tags import (
-    get_form_list, load_user_choice_poll_form)
+    get_form_list, url_to_anchor,
+    load_user_choice_poll_form, forms_list_linked_to_pages)
 from .base import create_form
 
 
@@ -296,6 +297,18 @@ class FormListTest(TestCase, MoloTestCaseMixin):
             self.translated_direct_form not in context['forms'])
         self.assertTrue(self.translated_linked_form in context['forms'])
 
+    def test_forms_list_linked_to_pages(self):
+        context = Context({
+            'locale_code': 'en',
+            'request': self.request,
+        })
+        survey = self.direct_molo_form_page
+        article = self.mk_article(self.main)
+        article_page_form = ArticlePageForms(form=survey, page=article)
+        article.forms.add(article_page_form)
+        res = forms_list_linked_to_pages(context, article)
+        self.assertEqual(res['forms'][0]['molo_form_page'], survey)
+
     def test_get_form_list_arg_error(self):
         context = Context({
             'locale_code': 'en',
@@ -325,3 +338,13 @@ class FormListTest(TestCase, MoloTestCaseMixin):
             self.translated_direct_form not in context['forms'])
         self.assertTrue(
             self.trans_personalisable_form in context['forms'])
+
+
+class TestUrlToAnchorTemplateFilter(TestCase):
+
+    def test_url_to_anchor(self):
+        link = 'http://abc.com?somerandomvar=123'
+        self.assertEqual(
+            '<a href="{0}">{0}</a>'.format(link),
+            url_to_anchor(link)
+        )